Add configure/makefile infrastructure to ship plugins
BitlBee supports plugins, but the only plugins we ship ourselves are
special and can also be built into the binary. We don't want to do that
for all plugins we may potentially ship, so add some infrastructure to
easily add plugins.

All you now need to do to add a plugin is drop a .c file in plugins/ and
maybe add a FLAGS_plugin_name.so setting to the plugins/Makefile. Users
can then opt to build/install your plugin at configure time.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>