Add a plugin that turns bitlbee into a specialized jabber proxy
I have a bitlbee instance that is to be used only to connect to the
corporate Jabber server. This plugin automates that connection for
users, building on the ldap authentication backend, locked
accounts/settings (and the mysql storage backend, but it works with xml
storage as well).

It automates the following:
- When authenticating and no accounts have been set up, set up a jabber
  account based on data from LDAP.
- Disable password storage and reuse the bitlbee password for jabber
- When joining a channel, assume it's a jabber room and autoconfigure
  it.