conf: support AccountStorage as well as Account_Storage
These variables added way back in b73ac9c3 on 2005-12-14 have possibly
never been used, but out of an abundance of caution let's support the
underbar versions.

The reason to support the CamelCase version is that I'm
forward-porting https://github.com/bitlbee/bitlbee/pull/67 there's
production code that uses the CamelCase variant, and it would be handy
if it worked.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>